requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
anthropic
faiss-cpu
pydantic
pgvector
sentence-transformers
trafilatura
streamlit
strands-agents[anthropic]
strands-agents-tools
pyyaml

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
